remove use of soledad_writer
Since the soledad client lock gets us covered with writes now, it makes
no sense to enqueue using the messageconsumer.

The SoledadWriter is left orphaned by now. We might want to reuse it
to enqueue low priority tasks that need a strategy of retries in case
of revisionconflicts.

the MessageConsumer abstraction should also be useful for the case of
the smtp queue.