Remove notify_new callbacks from fetch and copy.
This fixes a bug with qtreactor that was making the 'OK foo copied'
not being delivered.

This or something similar will probably have to be re-added, because
on the current state the destination folder will not receive the
notification if it's selected *before* the copy operation has finished.

But in this way we have a clean slate that is working properly.
The bottleneck in the copy/append operations seems to have moved to
the select operation now.