Fix module import for iterdump method (#2945).
The python `pysqlcipher.dbapi2.Connection.iterdump` method is bound to the
`pysqlite_connection_iterdump()` function in src/connection.c.

This function should import the python `pysqlcipher.dump` module, which
contains the function `_iterdump` with the actual implementation of the
`pysqlcipher.dbapi2.Connection.iterdump` method.

The `pysqlcipher` module was adapted from the `pysqlite2` module. When that
was done we forgot to change the import of the `dump` module, which until now
was being loaded from the old `pysqlite2` module, instead of being loaded
from the correct `pysqlcipher` module.

This was a bug and was forcing us to depend on `pysqlite2` to be able to use
the `iterdump` method.

This commit fixes that bug by loading the correct `pysqlcipher` module in the
`pysqlite_connection_iterdump()` function in `src/connection.c`.