[bug] add extra CRLF to avoid bad mime parsing in Thunderbird
Thunderbird (as of 37.0b1) will display a blank body (with no
attachments) if some conditions are met:

* disk synchronization is disabled
* mime_part_on_demand = true
* msg size is bigger than the parts_on_demand threshold (30000 by
  default).

Comparing the logs with a well behaved imap server (dovecot, on this
case), it's easy to see that twisted implementation is lacking an extra
line separator at the end of each group of headers that is rendered in
response to each of the `BODY.PEEK[X.MIME]` command that the
mime_parts_on_demand will issue after getting the BODYSTRUCTURE.

This change patches the spew_body command on the body server.
We still would have to see if this is a bad behaviour in the thunderbird
side. The most similar bug I've found is:

https://bugzilla.mozilla.org/show_bug.cgi?id=149771

Which apparently was happening with exchange server.

We should send the patch to upstream twisted as well.

Note that this fix is not enough: the following commit, about fixing the
case of the boundary passed in the BODYSTRUCTURE response is also needed
to fix the bug (since a bad parsing happens all the same).

Resolves: #6773, #5010
Documentation: #6773
Releases: 0.4.0