[bug] re-add fix for multiple headers
This fix stores as multi-line headers that are repeated, and that were
being discarded when storing them in a regular dict.

It had been removed during the last refactor.

I also store headers now as a case-insensitive dict, which solves other
problems with the implementation of the twisted imap.

Releases: 0.4.0