[feature] post-sync mail processing hooks
using the new soledad plugin capablity, mail hooks to the post-sync
event by subscribing to the Meta-Doc type of documents. In this way, we
can create the uid tables and the uid entries needed to keep local
indexes for mail that has been processed in another instance.

however, this won't prevent a conflict if a given mail is received and
processed in two different instances. that is a problem that we still
have to deal with.

Resolves: #6996
Releases: 0.4.0