[bug] pass the doclist to the ui signal
before, we were taking the length of a string, signalling an incorrect
number to the ui.

currently this event is not being used, just only logged. in the future
the ui could probably might want to make use of this info to keep record
of a separate counter (how many mails received in the last sync).