[bug] saving message to drafts folder hangs
the bug consist on a fetch-while-pending-inserts hanging. the pending
insert dict was not being cleaned up because the lookup for the
Message-Id *is* case-sensitive (in the headers dict).

by using a temporary all-keys-lowercase dict the lookup can be performed
right, and the fetch returns successfully.

at this point there's still a pending bug with Drafts, and it is that
the new version is inserted but the MUA (TB) doesn't hide the older
version (although a Delete flag is added).

Resolves: #7189, #7190
Releases: 0.4.0