[bug] Fix missing _normailize_dict in DocumentWrapper constructor.
In the constructor values already is normalized (i.e. with underscores),
while kwargs contains items that are not normalized (i.e. with dashes).
Joining the dicts resulted in two entries that only differed by dash or
underscores. The setattr then set the value that occurred later in
items, thereby sometimes overriding the correct value with the default
one.