[feature] improve getmail utility
So now it:

- Accepts credentials in a file pointed by environment variable.
- Allows to specify the mailbox to select as a command line flag.
- Allows to select a given message by subject.

For example:

BITMASK_CREDENTIALS=/tmp/bm.secrets ./getmail --mailbox INBOX --subject
'test mail

The two flags are case-insensitive.

This is intended to be used as a helper in end-to-end tests.
Getting a message by subject it's suboptimal, but I think it's good
enough for our testing purposes right now.

Related: #7427