[bug] let the inbox used in IncomingMail notify any subscribed Mailbox
the mail service uses an Account object created from scratch, so it
wasn't sharing the collections mapping with the other Account object
that is created in the IMAP Service.

I make it a class attribute to allow mailbox notifications.
However, with the transition to a single service tree, this class
attribute can again become a class instance.

This is somehow related to a PR proposed recently by cz8s in pixelated
team:
https://github.com/leapcode/leap_mail/pull/228

However, I'm reluctant to re-use IMAPMailbox instances, since they
represent concurrent views over the same collection. I believe that
sharing the same underlying collection might be enough.