[bug] Adds user_id to Account
Previously Account used user id from the store, but this attribute is
optional and None by default. This caused the collection_mapping to be
unable to distinct between multiple users message collections.

This chance adds a non optional user_id attribute to Account and use it
to index the collection_mapping.

- Resolves: https://github.com/pixelated/pixelated-user-agent/issues/674
- Releases: 0.4.0