[tests] add test to validate signature from apple mail
This change adds test to validate signature of encrypted email created
by apple mail. It's important to note that apple mail has a specific way
to encrypt signed messages. First it sign the email and then encrypt the
previous signed message. It was also added a message file with the
expected data.