[refactor] change the check to validate signature from Apple Mail
It changes the way that incoming service checks if a additional
verification is needed to validate signature. The way before was
checking by the type of signature object and calling the verify
signature method if the type is different from OpenPGPKey. However it
could be more readable if we check the type of decrypted message. If
it's a multipart/signed message and not a plain/text we need to verify
the signature because keymanager couldn't do it during the decryption
process.