[bug] return logout message got from bonafide (fixes #8461)
The CLI prints results of commands using a default dictionary printer.
If the command result is not a dictionary, that writer will fail. The
logout command was returning None, so the printer was breaking when
using the CLI to logout.

This commit makes the logout command actually return the dictionary with
the logout message, thus not breaking the printer.