[bug] workaround around scrypt.encrypt() failure in windows
scrypt raises a "unable to get salt from /dev/urandom" error when using
encrypt method in windows.

this is a quick workaround for obfuscating the uuid map, but it needs to
be carefully reviewed as per the security implications (i'm using the
hostname to salt the key hashing).

doing this, I contemplated the possibility that the scrypt encryption we
were using could be flawed, since the payloads can be easily guessed.

so, take this with a pinch of salt (haha), and let's review it for the
future. as always, this was in a rush trying to bring windows support.