[feat] fallback support for PySide
basically, I give up trying to compile pyqt5 for python2. not worthy.
fortunately, the api changes are minimal so we can support two.

the way forward should be Py3.

still some bugs with rendering the viewport.
trying to help that, I introduced the web inspector, use DEBUG=1 flag to
have it visible.