[feature] serve UI as an hidden service
you need to configure onion = True in the services config.

for sure, this is gonna be more interesting when we ship pixelated.
but for now I thought it can be handy for testing ui changes: no need to
bundle, just use your tor browser :)

still need to try to serve imap/smtp over the onion service.

kudos to meejah for this super-usable library!