[bug] add tests for outgoing service
This commit originall contained a fix for a wrong variable name that
avoided the outgoing service to be launched. This had been fixed in a
concurrent mr that got merged before this one. However, this MR
contained a couple of tests for the OutgoingService, and I'm merging
those tests now.