[tests] add unit tests for the keymanager token fix
my goal here is also to start *real* unit tests for the keymanager, that
do not need an actual soledad implementation. this should make all the
coming tests much faster. I would like to have a common soledad stub
that follows the soledad api, but w/o actual disk persistance, and
migrate all the bulk of the keymanager tests to these unit tests.