[bug] fix the logout call
if user attempts to logout before the incoming multiservice has an
entry, there will be a KeyError raised on the
MultiService.getServiceNamed() call, which is improperly reported as a
confusing error message in the api return call.

by catching the KeyError, we make sure that the logout call can
terminate properly.